+++ source/Application/Controller/SuggestController.php
@@ -6,11 +6,11 @@
 
 namespace OxidEsales\EshopCommunity\Application\Controller;
 
-use oxRegistry;
+use OxidEsales\Eshop\Core\Registry;
 
 /**
  * Article suggestion page.
- * Collects some article base information, sets default recomendation text,
+ * Collects some article base information, sets default recommendation text,
  * sends suggestion mail to user.
  */
 class SuggestController extends \OxidEsales\Eshop\Application\Controller\FrontendController
@@ -60,6 +60,15 @@ class SuggestController extends \OxidEsales\Eshop\Application\Controller\Fronten
      * @var object
      */
     protected $_aSuggestData = null;
+
+    /**
+     * Assures, that controller would not be accessed if functionality disabled.
+     */
+    public function init()
+    {
+        $this->redirectToHomeIfDisabled();
+        parent::init();
+    }
 
     /**
      * Sends product suggestion mail and returns a URL according to
@@ -273,5 +282,15 @@ class SuggestController extends \OxidEsales\Eshop\Application\Controller\Fronten
         $aPaths[] = $aPath;
 
         return $aPaths;
+    }
+
+    /**
+     * In case functionality disabled, redirects to home page.
+     */
+    private function redirectToHomeIfDisabled()
+    {
+        if ($this->getConfig()->getConfigParam('blAllowSuggestArticle') !== true) {
+            Registry::getUtils()->redirect($this->getConfig()->getShopHomeUrl(), true, 301);
+        }
     }
 }