+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Possibility to delete user account via new methods:
   - `OxidEsales\EshopCommunity\Application\Controller\AccountController::deleteAccount()`
   - `OxidEsales\EshopCommunity\Application\Controller\AccountController::isUserAllowedToDeleteOwnAccount()`
+  - `OxidEsales\EshopCommunity\Core\ViewConfig::getShowSuggest()`
+- New settings:
+  - `blAllowSuggestArticle` - now it's possible to disable recommendation feature in admin.
 
 ### Changed